+++ CHANGES.md
@@ -14,15 +14,22 @@
 
 - analysis.plot2Dnet() 'view' param to select frontal ('xy') or top-down ('xz') view 
 
-- analysis.plotConn() option to plot stacked bar graph and select different pre and post pops
+- analysis.plotConn() option to plot stacked bar graph, select different pre and post pops, synMech and syns vs conns
 
-- Made distinction between number of connections and synaptic contacts (eg. can have 5 synsPerconn)
+- Made distinction between number of connections and synaptic contacts (eg. can have 5 synsPerConn)
+
+- New function netParams.addCellParamsSecList to create list of sections based on distance from soma
 
 - Removed 'plast' and 'shape' from conns when not used (issue #177)
+
+- Added option to center soma at origin when importing cell 
 
 - Now using noiseFromRandom123() for NetStims() (issue #89)
 
 - Fixed bug in readCmdLineArgs() 
+
+- Fixed bug in saveData option of figure plots
+
 
 
 # Version 0.6.4